+++ R/languages.R
@@ -3,5 +3,12 @@
 #' Lists the languages available for MDT implementations.
 #' @export
 mdt_languages <- function() {
-  c("EN", "DE", "FR", "RU")
+  c(
+    English = "en",
+    French = "fr",
+    `German (informal)` = "de",
+    `German (formal)` = "de_f",
+    Italian = "it",
+    Russian = "ru"
+  )
 }