+++ data-raw/mdt-dict.R
@@ -7,5 +7,17 @@ stopifnot(!anyDuplicated(input$key),
           all(input$key == russian$key))
 input$RU <- russian$RU
 
+italian <- read.csv("data-raw/dict-italian.csv", stringsAsFactors = FALSE)
+names(italian)[[1]] <- "key"
+stopifnot(!anyDuplicated(input$key),
+          all(input$key == italian$key))
+input$IT <- italian$IT
+
+german_formal <- read.csv("data-raw/dict-german-formal.csv", stringsAsFactors = FALSE)
+names(german_formal)[[1]] <- "key"
+stopifnot(!anyDuplicated(input$key),
+          all(input$key == german_formal$key))
+input$DE_F <- german_formal$DE_F
+
 mdt_dict <- psychTestR::i18n_dict$new(input)
 usethis::use_data(mdt_dict, overwrite = TRUE)